+++ mock.go
@@ -7,15 +7,17 @@ import (
 
 // MockClient is a mock Store implementation.
 type MockClient struct {
-	Hits   []Hit
-	Events []Event
-	m      sync.Mutex
+	Hits          []Hit
+	Events        []Event
+	ReturnSession *Session
+	m             sync.Mutex
 }
 
 // NewMockClient returns a new mock client.
 func NewMockClient() *MockClient {
 	return &MockClient{
-		Hits: make([]Hit, 0),
+		Hits:   make([]Hit, 0),
+		Events: make([]Event, 0),
 	}
 }
 
@@ -36,8 +38,16 @@ func (client *MockClient) SaveEvents(events []Event) error {
 }
 
 // Session implements the Store interface.
-func (client *MockClient) Session(clientID int64, fingerprint string, maxAge time.Time) (string, time.Time, time.Time, error) {
-	return "", time.Now().UTC(), time.Now().UTC(), nil
+func (client *MockClient) Session(clientID int64, fingerprint string, maxAge time.Time) (Session, error) {
+	if client.ReturnSession != nil {
+		return *client.ReturnSession, nil
+	}
+
+	return Session{
+		Path:    "",
+		Time:    time.Now().UTC(),
+		Session: time.Now().UTC(),
+	}, nil
 }
 
 // Count implements the Store interface.