+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.6.0
+
+* added cache to read sessions
+* added filter inversion
+* fixed session query order by
+* updated dependencies
+
 ## 2.5.2
 
 * fixed usage of TrackerConfig.SessionMaxAge